unify date format to Month DD, YYYY + dedupe bookmarks at same spot
Date format: every client-side date renderer was using
month:'short' (e.g. "Apr 18, 2026"), which clashed with the
canonical Month DD, YYYY format the readableDate / shortDate
filters produce server-side. Switched all five sites to
month:'long' so a date never reads "Apr 18" in one place and
"April 18" in another:

  annotations.js:156          (annotation list)
  library.js:338              (library annotation list)
  notes-page.js:80            (notes page activity dates)
  revision-history.js:31      (revision-history sidebar dates)
  archives-timeline.njk:514   (archives timeline; also added
                               year for full canonical format)

Bookmarks: the in-page bookmark button (.ann-bookmark-btn) used
to push a new entry on every click, so a reader who tapped twice
or held the bar at the same scroll position ended up with three
identical pins. Both bookmark stores now skip the add when an
existing entry sits at the same location:

  Articles (annotations.js Bookmarks)   skip when scrollPct is
                                        within 2 percentage points
                                        OR bodyOffset is within
                                        60px of an existing pin
  Library  (library.js Bookmarks)       skip when scrollPct is
                                        within 2 percentage points
                                        of an existing pin in the
                                        same chapter

The "saved" flash on the button still fires either way — a
duplicate tap reads as success because the bookmark is at the
spot the user wanted, we just don't redundantly persist it.

+++ src/assets/js/notes-page.js
@@ -77,7 +77,7 @@ function boot(root) {
   function formatDate(ts) {
     try {
       var d = new Date(ts);
-      return d.toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' });
+      return d.toLocaleDateString(undefined, { year: 'numeric', month: 'long', day: 'numeric' });
     } catch (e) { return ''; }
   }
