+++ src/assets/js/revision-history.js
@@ -28,7 +28,7 @@
   function formatDate(iso) {
     try {
       var d = new Date(iso);
-      return d.toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' });
+      return d.toLocaleDateString(undefined, { year: 'numeric', month: 'long', day: 'numeric' });
     } catch (e) { return iso; }
   }
